requirements.txt
streamlit
numpy
opencv-python-headless
Pillow
pdf2image
pytesseract